+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.IntegrationTests/DomainTests/ChargeLinks/MessageHub/ChargeLinkDataAvailableNotifierEndpointTests.cs
@@ -29,7 +29,7 @@ using Xunit;
 using Xunit.Abstractions;
 using Xunit.Categories;
 
-namespace GreenEnergyHub.Charges.IntegrationTests.Functions.ChargeLinks.MessageHub
+namespace GreenEnergyHub.Charges.IntegrationTests.DomainTests.ChargeLinks.MessageHub
 {
     [IntegrationTest]
     public class ChargeLinkDataAvailableNotifierEndpointTests
