+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.IntegrationTests/DomainTests/Charges/ChargeIngestionTests.cs
@@ -28,7 +28,7 @@ using Xunit;
 using Xunit.Abstractions;
 using Xunit.Categories;
 
-namespace GreenEnergyHub.Charges.IntegrationTests.Functions.Charges
+namespace GreenEnergyHub.Charges.IntegrationTests.DomainTests.Charges
 {
     /// <summary>
     /// Proof-of-concept on integration testing a function.
